requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.0
dateparser>=1.2.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
